aggregation_service/requirements.txt
pymongo
pyzmq
opencv-python-headless
